Moved EditmarkupLink component to Components
